requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.2.0
ffmpeg-python==0.2.0
frozenlist==1.4.0
future==0.18.3
idna==3.4
imageio-ffmpeg==0.4.8
multidict==6.0.4
numpy==1.25.2
PlexAPI==4.15.1
py-cord==2.4.1
python-dotenv==1.0.0
requests==2.31.0
typing_extensions==4.7.1
urllib3==2.0.4
yarl==1.9.2
